:book: API Documentation Work

+++ bindings/timers/ClearTimeout.go
@@ -4,7 +4,7 @@ package timers
 
 import "syscall/js"
 
-func ClearTimeout(handler_id uint) {
+func ClearTimeout(handler_id uint64) {
 
 	wrapped_identifier := js.ValueOf(handler_id)
